requirements.txt
fastapi[standard]
jinja2
python-multipart
email-validator
python-dotenv
uvicorn
pydantic[email]